inline sparklines in xunit
