+++ allure-report-face/package.json
@@ -66,6 +66,7 @@
     "i18next": "^2.1.0",
     "i18next-intervalplural-postprocessor": "0.0.8",
     "jquery": "^2.1.4",
+    "jquery-sparkline": "^2.3.2",
     "normalize-css": "^2.3.1",
     "sortablejs": "^1.4.2",
     "underscore": "^1.8.3"
