+++ allure-report-face/src/components/pane-set/styles.css
@@ -63,3 +63,7 @@
     }
   }
 }
+
+.pane + .pane {
+  box-shadow: 0 0 3px #eee;
+}
