+++ allure-report-face/src/plugins/xunit/testsuites-list/TestsuitesListView.hbs
@@ -1,17 +1,17 @@
-<a class="pane__overlay" href="#{{baseUrl}}"></a>
-<h2 class="pane__title {{b "testsuites-list" "title"}}">
+<a class="{{b 'pane' 'overlay'}}" href="#{{baseUrl}}"></a>
+<h2 class="{{b 'pane' 'title'}} {{b "testsuites-list" "title"}}">
     xUnit results
-    <span class="testsuites-list__statuses"></span>
+    <span class="{{b "testsuites-list" "statuses"}}"></span>
 </h2>
 
-<div class="{{b "table" hover=true}} testsuites-list__table">
+<div class="{{b "table" hover=true}} {{b 'testsuites-list' 'table'}}">
     <div class="{{b "table" "head"}}">
         {{> ../../../blocks/table/sortable-col field="title" title="Title" class=(b "testsuites-list" "name")}}
         {{> ../../../blocks/table/sortable-col field="time.duration" title="Duration"}}
-        {{> ../../../blocks/table/sortable-col field="statistic.failed" title="Failed"}}
-        {{> ../../../blocks/table/sortable-col field="statistic.broken" title="Broken"}}
-        {{> ../../../blocks/table/sortable-col field="statistic.canceled" title="Canceled"}}
-        {{> ../../../blocks/table/sortable-col field="statistic.pending" title="Pending"}}
+        {{> ../../../blocks/table/sortable-col field="statistic.failed" class=(b "table" "col" failed=true)  title="Failed"}}
+        {{> ../../../blocks/table/sortable-col field="statistic.broken" class=(b "table" "col" broken=true) title="Broken"}}
+        {{> ../../../blocks/table/sortable-col field="statistic.canceled" class=(b "table" "col" canceled=true) title="Canceled"}}
+        {{> ../../../blocks/table/sortable-col field="statistic.pending" class=(b "table" "col" pending=true) title="Pending"}}
         {{> ../../../blocks/table/sortable-col field="statistic.total" title="Total"}}
     </div>
     <div class="{{b "table" "row" summary=true}}" disabled>
@@ -23,7 +23,7 @@
     {{#each suites}}
         <a class="{{b "table" "row"}}" data-uid="{{uid}}" href="#{{../baseUrl}}/{{uid}}">
             <div class="{{b "table" "col"}} {{b "testsuites-list" "name"}} long-line">
-                {{title}}
+                <span class="inlinesparkline">{{statistic.failed}},{{statistic.broken}},{{statistic.canceled}},{{statistic.pending}},{{statistic.passed}}</span> {{title}}  
             </div>
             {{> suite-table-row}}
         </a>
