+++ allure-report-face/src/plugins/xunit/testsuites-list/TestsuitesListView.js
@@ -5,6 +5,8 @@ import StatusToggleView from '../../../components/status-toggle/StatusToggleView
 import {region} from '../../../decorators';
 import settings from '../../../util/settings';
 import template from './TestsuitesListView.hbs';
+import sparkline from 'jquery-sparkline'
+import {colors} from '../../../util/statuses'
 
 class TestsuitesListView extends DataGridView {
     template = template;
@@ -17,6 +19,10 @@ class TestsuitesListView extends DataGridView {
         this.state = state;
         this.listenTo(this.state, 'change:testsuite', (m, suite) => this.highlightItem(suite));
         this.listenTo(settings, 'change:visibleStatuses', this.render);
+    }
+
+    onDomRefresh() {
+        this.$el.find('.inlinesparkline').sparkline('html', {type: 'bar', colorMap: colors});
     }
 
     onRender() {
